Reorganize files, update Rakefile and Gemspec
