--- ConnectSDK/Gemfile
@@ -1,7 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in ConnectSDK.gemspec
-gemspec
-
-gem 'json'
-gem 'rake'
