--- ConnectSDK/lib/ConnectSDK/version.rb
@@ -1,3 +1,0 @@
-module ConnectSDK
-	VERSION = "1.0.2.Beta"
-end
