+++ Gemfile
@@ -1,6 +1,5 @@
 source 'https://rubygems.org'
 
-gem 'ConnectSDK'
 gem 'rake'
 gem 'json'
 gem 'test-unit'
