+++ Rakefile
@@ -1,2 +1,10 @@
+# require "bundler"
 require "bundler/gem_tasks"
 
+require "rake/testtask"
+
+task :test do
+  ruby("unit_tests/Run.rb")
+end
+
+task default: :test
