+++ lib/ConnectSDK/version.rb
@@ -1,0 +1,3 @@
+module ConnectSDK
+	VERSION = "2.0.0"
+end
