+++ lib/Images/Images.rb
