+++ lib/Search/SearchImages.rb
