+++ lib/Search/SearchImagesEditorial.rb
