+++ lib/Search/SearchVideosEditorial.rb
