+++ unit_tests/CustomRequestsTests.rb
@@ -1,6 +1,6 @@
 require 'test/unit'
 require 'webmock/test_unit'
-require_relative "../ConnectSDK/lib/ConnectSDK.rb"
+require_relative "../lib/ConnectSDK.rb"
 
  
 class CustomRequestsTests < Test::Unit::TestCase
