+++ unit_tests/DownloadVideosTests.rb
@@ -1,6 +1,6 @@
 require 'test/unit'
 require 'webmock/test_unit'
-require_relative "../ConnectSDK/lib/ConnectSDK.rb"
+require_relative "../lib/ConnectSDK.rb"
 
  
 class DownloadVideosTests < Test::Unit::TestCase
